pyproject.toml
[tool.poetry]
name = "fireemblem-ctl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyAutoGUI = "^0.9.53"
Pillow = "^9.1.0"
pytesseract = "^0.3.9"
opencv-python = "^4.5.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
nose2 = "^0.11.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
